Separate tests that rely on SecurityManager (JLM_Tests)
SecurityManager will be removed in a future jdk. Separate tests that
rely on it so that they can be easily disabled in the future.

Issue: #14412
