+++ test/functional/JLM_Tests/testng.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <!--
-  Copyright (c) 2016, 2019 IBM Corp. and others
+  Copyright (c) 2016, 2022 IBM Corp. and others
 
   This program and the accompanying materials are made available under
   the terms of the Eclipse Public License 2.0 which accompanies this
@@ -90,10 +90,12 @@
 			<class name="org.openj9.test.java.lang.management.TestLoggingMXBean" />
 			<!-- //not run in rtctest <class name="org.openj9.test.java.lang.management.TestJvmCpuMonitorMXBeanEx"/> -->
 			<!-- // This test runs fine when executed standalone, but when // executed
-				in batch (with other tests), it fails. <class name="org.openj9.test.java.lang.management.TestSharedClassMemoryMXBean"/> -->
+				in batch (with other tests), it fails. <class name="org.openj9.test.java.lang.management.TestSharedClassMemoryMXBean"/>
+				SecurityManager tests: <class name="org.openj9.test.java.lang.management.TestSharedClassMemoryMXBean_SM"/> -->
 			<!-- // Fixing this test suite is out of scope - neither are these //
 				tests related to DynamicMBean V. StandardMBean nor do they // seem to have
-				simple fixes. <class name="org.openj9.test.java.lang.management.TestRegression"/> -->
+				simple fixes. <class name="org.openj9.test.java.lang.management.TestRegression"/>
+				SecurityManager tests: <class name="org.openj9.test.java.lang.management.TestRegression_SM"/> -->
 		</classes>
 	</test> <!-- JLM_Tests -->
